requirements.txt
asgiref==3.3.4
Django==3.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
djangorestframework==3.12.4
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
